#2231 - Add support for WADO / Streaming
Interim commit.

Initial working version of an image streaming server based on WADO. Only application/dicom mime type is being supported.

- Added HttpListenerShred as one of the shred types that can be launched. The image server is launched as a http listener shred.
- Added ImageStreamingServer, renamed StreamingServer to HeaderStreamingServer.
- Added WADORequestProcessor to process WADO http requests intercept by the image streaming server. Currently it only supports application/dicom mime type. New mime type handler plugins can be added later.
- Updated store procedure to return study location based on only study instance uid. There's no concept of a server partition in WADO.


Other changes:
- Updated ShredController to set the name of the thread that executes the shred based on the shred name instead of "Shred N" (N=1,2,3..). This makes it easier to identify shred in the log.
- Added test client to retrieve images from the image streaming server.
- Fixed minor bug in AverageByteCountStatistics class.
- Modified TimeSpanStatistics and RateStatistics classes to allow access to the elapse time and current speed at anytime instead of having to wait until End() is called. This will allow us to display the "live" transmission speed.

+++ Common/Statistics/AverageByteCountStatistics.cs
@@ -87,12 +87,12 @@ namespace ClearCanvas.Common.Statistics
             }
             else if (sample is long)
             {
-                Samples.Add((ulong) (object) sample);
+                Samples.Add( (ulong) (long) (object) sample);
                 NewSamepleAdded = true;
             }
             else if (sample is int)
             {
-                Samples.Add((ulong) (object) sample);
+                Samples.Add((ulong) (int) (object) sample);
                 NewSamepleAdded = true;
             }
             else if (sample is uint)
